Winery Class Overhaul = Sexy
Rewrote most of the class for controlling the Winery list data.

Added several methods including a few methods for getting the name of a
specific winery as well as getting the distance between two wineries.
Other methods like getting the total number of wineries as well.

changed the file so that the number of the winery holding its own data
is no longer "0.0" but corresponds to the distance to Canyon Villa in
miles.