Working: AddWinery, AddWine
Both of these methods are working and (for the most part) tested well.
When the time comes I will add the functionality to the "AddWinery"
method that will allow it to modify the file.

I am only leaving that out for now because I don't want it to cause
problems to the file while we are testing the program